[DEBUG] Reservation history code structure fixed
Reservation history fixed for both admin and user
